services/infra-invitation-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
pydantic[email]==2.6.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
redis==5.0.1
aio-pika==9.4.0
